requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.6
Django==6.0.3
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.1.1
PyJWT==2.12.1
python-dotenv==1.2.2
qrcode==8.2
reportlab==4.4.10
sqlparse==0.5.5